refactor batch engine file layout and docs
Move batch-only internals under core/batching to make the batch engine easier to navigate, and document Editor.withBatch plus Transforms.applyBatch in the public docs.
