+++ packages/slate/src/core/batching/direct-text-merge-batch-children.ts
@@ -1,6 +1,6 @@
-import { Descendant, Element, Text } from '..'
-import { BaseMergeNodeOperation } from '../interfaces/operation'
-import { replaceChildren } from '../utils/modify'
+import { Descendant, Element, Text } from '../..'
+import { BaseMergeNodeOperation } from '../../interfaces/operation'
+import { replaceChildren } from '../../utils/modify'
 
 export const applyDirectTextMergeBatchToChildren = (
   children: Descendant[],
